backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
pillow>=10.2.0
numpy>=1.26.0
python-magic-bin>=0.4.14; sys_platform == "win32"
python-magic>=0.4.27; sys_platform != "win32"
python-dotenv>=1.0.0

# --- Database ---
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0    # Postgres driver (no-op for SQLite local dev)

# --- Cloud storage ---
cloudinary>=1.40.0

# --- Auth ---
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
passlib[bcrypt]>=1.7.4
google-auth>=2.28.0

# --- ML / detection ---
# Comment out the 3 lines below if deploying to a free tier (low RAM <1GB):
mediapipe>=0.10.18
ultralytics>=8.3.0
opencv-python-headless>=4.9.0
